atelier-ai/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pillow==10.2.0
python-multipart==0.0.6
numpy==1.26.3
scipy==1.11.4
# Add torch/torchvision here if you are installing the real model
# torch --index-url https://download.pytorch.org/whl/cu118
